Nest: Add net_peer route type to track discovered peers
The definition and helper functions for a new route-like
object to track peers discovery data has been added.
It only contains the (v4 or v6) remote peer address and the
ingress iface index, for now.

The main intent of this is, currently, to enable BGP
unnumbered auto peer discovery via RAdv incoming
advertisments, but in the future the same data structure
could be used to allow discovery coming from different protocols.

Signed-off-by: Matteo Perin <matteo.perin@canonical.com>