BGP: Add peers channel and sessions spawn on export to achieve automatic peering
This change proposal is made in order to achieve BGP automatic peering based
on discovered link-local addresses leveraging the IPv6 ND protocol.

A big part of the necessary logic is already present in the dynamic BGP implementation.
The problem is that dynamic BGP works in a passive/reactive model only, as it waits for
peers within a specified range to connect.

This integrates with this implementation by triggering the spawn of a BGP process
sending OPEN requests in order to achieve active peer connection when a valid peer
remote address has been discovered.

The discovery is done by exporting routing information from a table containing
peer data through a newly introduced peers channel. This means that the feature
will only work if there is another protocol actively importing peer data (e.g. RAdv).
The feature is also meant to work only in the presence of a valid dynamic BGP
configuration.

A new configuration paramenter has been added to the BGP protocol, persist [yes/no]
which controls the behavior of sessions shutdown on peer route withdrawal.
If enabled, a BGP connection spawned from peer discovery or dyamic BGP will be left running.
If disabled, the session will be torn down and removed.

Logic to avoid duplicate sessions from both the dynamic BGP and automatic peers
discovery paths has also been introduced.

Signed-off-by: Matteo Perin <matteo.perin@canonical.com>